% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af-ctrs.R
\name{afib_score}
\alias{afib_score}
\title{Atrial Fibrillation Composite Treatment Response Score}
\usage{
afib_score(
  symptom_type = c("AFEQT", "NYHA"),
  afeqt_total = NULL,
  afeqt_symptoms = NULL,
  afeqt_activities = NULL,
  afeqt_treatment = NULL,
  nyha_class = NULL,
  electrical_type = c("ECG", "MCT"),
  mct_burden = NULL,
  ecg_total = NULL,
  ecg_sinus = NULL,
  ecg_time_range = NULL
)
}
\arguments{
\item{symptom_type}{A \code{character} string indicating the type of symptom
burden ("AFEQT" or "NYHA").}

\item{afeqt_total}{A \code{numeric} value representing the total AFEQT score
(0-100), if applicable.}

\item{afeqt_symptoms}{A \code{numeric} value representing the AFEQT symptoms
domain score (0-100), if applicable.}

\item{afeqt_activities}{A \code{numeric} value representing the AFEQT daily
activities domain score (0-100), if applicable.}

\item{afeqt_treatment}{A \code{numeric} value representing the AFEQT treatment
concerns domain score (0-100), if applicable.}

\item{nyha_class}{A \code{character} string representing the NYHA class ("I",
"II", "III", "IV"), if applicable.}

\item{electrical_type}{A \code{character} string indicating the type of data used
for electrical burden ("ECG" or "MCT").}

\item{mct_burden}{A \code{numeric} value representing the percentage of time in AF
from MCT data, if applicable.}

\item{ecg_total}{A \code{integer} value representing the total number of ECGs, if
applicable.}

\item{ecg_sinus}{A \code{integer} value representing the number of ECGs in sinus
rhythm, if applicable.}

\item{ecg_time_range}{A \code{character} string representing the time range for
ECG data (e.g., "6 months", "12 months").}
}
\value{
A \code{numeric} value representing the composite AF score.
}
\description{
This function calculates a composite score based on symptom burden and
electrical burden.
}
\examples{
afib_score(
  symptom_type = "AFEQT", afeqt_total = 80, afeqt_symptoms = 70,
  afeqt_activities = 85, afeqt_treatment = 75,
  electrical_type = "MCT", mct_burden = 0.2
)

afib_score(
  symptom_type = "NYHA", nyha_class = "III", electrical_type =
    "ECG", ecg_total = 10, ecg_sinus = 7, ecg_time_range = "12 months"
)

}
