% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-marksman.R
\name{aim}
\alias{aim}
\title{Aiming For Target Hypothesis}
\usage{
aim(bullets, ...)
}
\arguments{
\item{bullets}{List object of bullets that are to be thought of as the
statistical plan, as defined by each individual bullet. The list can be
named to make it simpler.}

\item{...}{For extensibility}
}
\value{
List of \code{aims} that contain all the relevant hypothesis that were
pre-specified, including workflows and model specifications from
\code{tidymodels}
}
\description{
This creates the hypothesis that are being prepared for
analysis. These contain model specifications and the workflows, along with
model identifiers.
}
\examples{
data(geh)
f <- svg_mag + qrs_tang ~ lab_hba1c + age + sex + bmi + cad + htn
lm_mod <- parsnip::linear_reg() \%>\% parsnip::set_engine("lm")
b1 <- bullet(f, exposure = "lab_hba1c", approach = "sequential", model = lm_mod)
b2 <- bullet(f, exposure = "lab_hba1c", approach = "parallel", model = lm_mod)
bullets <- list(seq = b1, par = b2)
}
