% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-methods.R
\name{anova_cosinor_models}
\alias{anova_cosinor_models}
\title{Compare Nested Cosinor Models}
\usage{
anova_cosinor_models(models)
}
\arguments{
\item{models}{a list of models of class \code{cosinor}}
}
\value{
a \code{data.frame} of class \code{anova}
}
\description{
Sequential F tests across a list of \code{cosinor} models fitted to
the same observations.
}
\keyword{internal}
