% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{approxExtrap}
\alias{approxExtrap}
\title{approxExtrap}
\usage{
approxExtrap(
  x,
  y,
  xout,
  method = "linear",
  n = 50,
  rule = 2,
  f = 0,
  ties = "ordered",
  na.rm = FALSE
)
}
\arguments{
\item{Q}{discharge}
}
\description{
description
}
\seealso{
ref
}
