% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-constructor.R
\name{augment.cosinor}
\alias{augment.cosinor}
\title{Augment data with information from a \code{cosinor} object}
\usage{
\method{augment}{cosinor}(x, ...)
}
\arguments{
\item{x}{A \code{cosinor} object created by \code{\link[=cosinor]{cosinor()}}}

\item{...}{For extensibility}
}
\value{
a \code{tibble} object
}
\description{
Augment accepts a \code{cosinor} model object and adds information about each observation in the dataset. This includes the predicted values in the \code{.fitted} column and the residuals in the \code{.resid} column. New columns always begin with a \code{.} prefix to avoid overwriting columns in original dataset.
}
\seealso{
Other cosinor:
\code{\link[=cosinor]{cosinor()}},
\code{\link[=ggcosinor]{ggcosinor()}}
}
\concept{cosinor}
