% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-marksman.R
\name{ballistics}
\alias{ballistics}
\alias{ballistics.default}
\alias{ballistics.bullet}
\alias{ballistics.aim}
\title{Ballistic Analysis of the Hypotheses}
\usage{
ballistics(mark, ...)

\method{ballistics}{default}(mark, ...)

\method{ballistics}{bullet}(mark, ...)

\method{ballistics}{aim}(mark, ...)
}
\arguments{
\item{mark}{Represents either a \code{bullet} or \code{aim} object to be analyzed for
further information or characteristics
\itemize{
\item A single \strong{bullet} object will return a table that is structured for the
proposed analytical approach
\item A single \strong{aim} object, after the \code{fire()} function, will return a table
that includes further analytical elements. Currently returns the runtime in
miliseconds.
}}

\item{...}{For extensibility}
}
\value{
The characteristics needed to create an appropriate modeling matrix when using the \code{aim()} function. Returns a tibble of the appropriate combination of outcomes/predictors.
}
\description{
Look at the overall characteristics of the hypotheses, such as
number of regression models and how they will be put together based on the
approach. Is used to evaluate the objects created by \code{bullet()} and by
\code{aim()} (which also uses ballistics to generate the appropriate modeling
matrix).
}
\examples{
data(geh)
f <- svg_mag + qrs_tang ~ lab_hba1c + age + sex + bmi + cad + htn
lm_mod <- parsnip::linear_reg() \%>\% parsnip::set_engine("lm")
b <- bullet(f, exposure = "lab_hba1c", approach = "parallel", model = lm_mod)
ballistics(b)
}
\concept{ballistics}
