% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrv_modeling.R
\name{build_models}
\alias{build_models}
\title{Model Building}
\usage{
build_models(formula, data, type, engine = "linear", exposure = NULL)
}
\arguments{
\item{formula}{an object of class \code{formula} that shows the names of the
outcomes (can be more than 1) and the names of the predictors (which should
contain the \code{exposure} variable). This formula can include mixed effects if
needed.}

\item{data}{data frame or data table (or tibble) that contains the named
variables}

\item{type}{Type of model that will be used. The options are:
\itemize{
\item \code{sequential} will build y ~ x1, y ~ x1 + x2 models
\item \code{parallel} will build y ~ x1, y ~ x2 models
}}

\item{engine}{Set the "engine" or the regression tool that will be used. One
day this may be a useful addition to the \code{tidymodels} approach, which is
the inspiration for the terminology. Currently limited to the following:
\itemize{
\item \code{linear} uses the stats::lm() for linear regression, and lme4::lmer() for
mixed effect models
}}

\item{exposure}{Variable that is forced to be maintained in every model as a
predictor. This is currently used only for \code{sequential} type models}
}
\value{
A tidy tibble of models. Each one will likely be grouped by its
outcome, and then with sequential columns using increased/additive models.
Each model, in a tidy format, will have two additional columns.
\itemize{
\item \code{outcomes} identifies which outcome was used for the specific regression
\item \code{covar} number of covariates used in sequence of predictors given, with
exposure always being placed in position 1
}
}
\description{
Simplify the process of building multiple models. Models are
usually built in sequential order, or in parallel when testing multiple
potential patterns. This is particularly helpful in epidemiological cases
of testing effect of additional parameters. Every parameter should be
theoretically a part of the causal model for the exposure-outcome
relationship.
}
\details{
This is considering what is available with the \code{modelr} package and
the \code{tidymodels} approach, and finding an in-between for the causality /
epidemiology approach of building intentional, sequentional models. Expect
changes in the process.
}
\examples{
data(geh)
f <- svg_mag + qrs_tang ~ lab_hba1c + bmi

}
