% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-marksman.R
\name{bullet}
\alias{bullet}
\title{Create bullets to be loaded}
\usage{
bullet(f, exposure = NULL, approach = "sequential", model, ...)
}
\arguments{
\item{f}{Formula showing relationship of outcomes to predictors}

\item{exposure}{Variable(s) that is forced to be maintained in every model as a
predictor.}

\item{approach}{The modeling approach that will be used. The options are:
\itemize{
\item \code{sequential} will build y ~ x1, y ~ x1 + x2 models
\item \code{parallel} will build y ~ x1, y ~ x2 models
}}

\item{model}{A model specified by the \href{https://parsnip.tidymodels.org/articles/parsnip_Intro.html}{\code{parsnip} package}, which
includes the mode and computational engine}

\item{...}{For extensibility}
}
\value{
List of hypothesis characteristics needed for analysis
}
\description{
Create bullets to be loaded
}
\examples{
data(geh)
f <- svg_mag + qrs_tang ~ lab_hba1c + age + sex + bmi + cad + htn
lm_mod <- parsnip::linear_reg() \%>\% parsnip::set_engine("lm")
b <- bullet(f, exposure = "lab_hba1c", approach = "sequential", model = lm_mod)
}
