% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CARD.visualize.gene}
\alias{CARD.visualize.gene}
\title{Visualize the spatial distribution of cell type proportion}
\usage{
CARD.visualize.gene(
  spatial_expression,
  spatial_location,
  gene.visualize,
  colors = colors,
  NumCols
)
}
\arguments{
\item{spatial_expression}{Data frame, spatial gene expression in either original resolution or enhanced resolution.}

\item{spatial_location}{Data frame, spatial location information.}

\item{gene.visualize}{Vector of selected gene names that are interested to visualize}

\item{colors}{Vector of color names that you want to use, if NULL, we will use the default color scale in virdis palette}

\item{NumCols}{Numeric, number of columns in the figure panel, it depends on the number of cell types you want to visualize.}
}
\value{
Returns a ggplot2 figure.
}
\description{
Visualize the spatial distribution of cell type proportion
}
