% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CARD.visualize.pie}
\alias{CARD.visualize.pie}
\title{Visualize the spatial distribution of cell type proportion in a geom scatterpie plot}
\usage{
CARD.visualize.pie(proportion, spatial_location, colors = colors)
}
\arguments{
\item{proportion}{Data frame, cell type proportion estimated by CARD in either original resolution or enhanced resolution.}

\item{spatial_location}{Data frame, spatial location information.}

\item{colors}{Vector of color names that you want to use, if NULL, we will use the default color scale c("lightblue","lightyellow","red")}
}
\value{
Returns a ggplot2 figure.
}
\description{
Visualize the spatial distribution of cell type proportion in a geom scatterpie plot
}
