% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CARD.visualize.prop.2CT}
\alias{CARD.visualize.prop.2CT}
\title{Visualize the spatial distribution of two cell type proportions on the same plot}
\usage{
CARD.visualize.prop.2CT(
  proportion,
  spatial_location,
  ct2.visualize = ct2.visualize,
  colors = NULL
)
}
\arguments{
\item{proportion}{Data frame, cell type proportion estimated by CARD in either original resolution or enhanced resolution.}

\item{spatial_location}{Data frame, spatial location information.}

\item{ct2.visualize}{Vector of selected two cell type names that are interested to visualize, here we only focus on two cell types}

\item{colors}{list of color names that you want to use for each cell type, if NULL, we will use the default color scale list
list(c("lightblue","lightyellow","red"),c("lightblue","lightyellow","black")}
}
\value{
Returns a ggplot2 figure.
}
\description{
Visualize the spatial distribution of two cell type proportions on the same plot
}
