% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD.prop.R
\name{CARD_deconvolution}
\alias{CARD_deconvolution}
\title{Spatially Informed Cell Type Deconvolution for Spatial Transcriptomics by CARD}
\usage{
CARD_deconvolution(CARD_object)
}
\arguments{
\item{CARD_object}{CARD object create by the createCARDObject function}
}
\value{
Returns a CARD object with estimated cell type proportion stored in object@Proportion_CARD.
}
\description{
Spatially Informed Cell Type Deconvolution for Spatial Transcriptomics by CARD
}
