% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_management.R
\name{CARD_list_all}
\alias{CARD_list_all}
\title{CARD_list_all}
\usage{
CARD_list_all(include_experimental = FALSE)
}
\description{
List all the CARD parametrization files in a \link[tibble:tibble]{tibble} to get all the metadata of the aggregation variables available.
}
\examples{
# Get all the available variables
metaEX_all = CARD_list_all()
# And why note filter it by topic
dplyr::filter(metaEX_all, grepl("Low Flow", topic_en))
}
\seealso{
\itemize{
\item 
\enumerate{
\item \code{\link[=CARD_list_all]{CARD_list_all()}} list all available CARD.
}
}
\enumerate{
\item \code{\link[=CARD_management]{CARD_management()}} for managing CARD parameterization files.
\item \code{\link[=CARD_extraction]{CARD_extraction()}} for extracting variables using CARD.
}
}
