% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD_management.R
\name{CARD_management}
\alias{CARD_management}
\title{CARD_management}
\usage{
CARD_management(
  CARD_name = c("QA", "QJXA"),
  CARD_path = "./WIP",
  add_id = TRUE,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{CARD_name}{A \link[base:c]{vector} of \link[base:character]{character} strings to specify which variables you want to extract. See \code{\link[=CARD_list_all]{CARD_list_all()}} to get the variable names. By default, \code{c("QA", "QJXA")}. If \code{NULL}, all the variable will be extracted, so avoid this value except with \code{extract_only_metadata = TRUE}.}

\item{CARD_path}{An optional \link[base:character]{character} string for the path where to search for custom CARDs that have been created by the \link{CARD_management} function. By default, \code{NULL} in order to get the default CARD variable parameters.}

\item{add_id}{\link[base:logical]{logical}. If \code{TRUE}, numerical IDs will be added to the start of the copied and pasted CARD names to maintain the input order. Default is \code{TRUE}.}

\item{overwrite}{\link[base:logical]{logical}. If \code{TRUE}, existing CARD files in the analysis directory will be overwritten. Default is \code{TRUE}.}

\item{verbose}{\link[base:logical]{logical}. Should intermediate messages be printed during the execution of the function ? Default \code{FALSE}.}
}
\value{
Selected CARD parameterization files will be created in the \code{CARD_path} directory.
}
\description{
Manage your different sets of variables to extract. For CARD advanced users, this function manages the CARD directory structure by performing automatic file operations to get the CARD parameterization files to your wanted directory in order to custom them or create yours to use them in a second step with \code{\link[=CARD_extraction]{CARD_extraction()}}. If you want to submit request for new CARD see the \href{https://github.com/super-lou/EXstat.CARD}{GitHub repo}\url{https://github.com/super-lou/EXstat.CARD}.
}
\examples{
# Get the QA and QMNA CARD variables in your local CARD_path directory
CARD_management(CARD_name=c("QA", "QMNA"), CARD_path="CARD-WIP")
}
\seealso{
\enumerate{
\item \code{\link[=CARD_list_all]{CARD_list_all()}} list all available CARD.
}
\itemize{
\item 
\enumerate{
\item \code{\link[=CARD_management]{CARD_management()}} for managing CARD parameterization files.
}
}
\enumerate{
\item \code{\link[=CARD_extraction]{CARD_extraction()}} for extracting variables using CARD.
}
}
