% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{class}
\name{CARDfree-class}
\alias{CARDfree-class}
\title{Each CARDfree object has a number of slots which store information. Key slots to access
are listed below.}
\description{
Each CARDfree object has a number of slots which store information. Key slots to access
are listed below.
}
\section{Slots}{

\describe{
\item{\code{spatial_countMat}}{The filtered spatial count data.}

\item{\code{spatial_location}}{The weights for combining p-values from multiple kernels.}

\item{\code{Proportion_CARD}}{The estimated cell type proportion by CARD with each row is a spatial location and each column is a cell type.}

\item{\code{estimated_refMatrix}}{The estimated reference matrix by CARDfree with each row represents a gene and each column represents a cell type cluster.}

\item{\code{project}}{The name of the project, default is deconvolution.}

\item{\code{markerList}}{The nlist of cell type specific markers, with each element represents the vector of cell type specific markers}

\item{\code{info_parameters}}{The paramters that are used in model fitting.}

\item{\code{algorithm_matrix}}{The intermediate matrices that are used in the model fitting step.}

\item{\code{refined_prop}}{The refined cell type proportion matrix estimated by CARD for the newly grided spatial locations. The number of initial grids are defined by the user.}

\item{\code{refined_expression}}{The refined predicted expression matrix (normalized) estimated by CARD for the newly grided spatial locations. The number of initial grids are defined by the user.}
}}

