% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CARDref}
\alias{CARDref}
\title{SpatialDeconv function based on Conditional Autoregressive model}
\usage{
CARDref(
  XinputIn,
  UIn,
  WIn,
  phiIn,
  max_iterIn,
  epsilonIn,
  initV,
  initb,
  initSigma_e2,
  initLambda
)
}
\arguments{
\item{XinputIn}{The input of normalized spatial data}

\item{UIn}{The input of cell type specific basis matrix B}

\item{WIn}{The constructed W weight matrix from Gaussian kernel}

\item{phiIn}{The phi value}

\item{max_iterIn}{Maximum iterations}

\item{epsilonIn}{epsilon for convergence}

\item{initV}{Initial matrix of cell type compositions V}

\item{initb}{Initial vector of cell type specific intercept}

\item{initSigma_e2}{Initial value of residual variance}

\item{initLambda}{Initial vector of cell type sepcific scalar.}
}
\value{
A list
}
\description{
SpatialDeconv function based on Conditional Autoregressive model
}
