% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian-patterns.R
\name{circ_rad}
\alias{circ_rad}
\title{Convert Time to Radians}
\usage{
circ_rad(t, period, ...)
}
\arguments{
\item{t}{vector of time units}

\item{period}{period of the time units (e.g. 24 hours)}
}
\value{
Vector converted into radians
}
\description{
Converts a time unit to radians given a known period. Allows for vectorization.
}
\details{
This function supports other centering functions to study time series.
}
