% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian_plots.R
\name{circadian_forest}
\alias{circadian_forest}
\title{Forest Plot of Hourly Odds}
\usage{
circadian_forest()
}
\description{
`circadian_forest` creates an OR table for each hour of data given
}
\details{
This function creates an OR table based on the covariate names supplied. It requires that there is an appropriate outcome variable selected.
}
