% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian_plots.R
\name{circadian_ttest}
\alias{circadian_ttest}
\title{T-test Table by Time Point}
\usage{
circadian_ttest(data, time, cvar, group)
}
\arguments{
\item{data}{Dataframe containing all the following variables}

\item{time}{Name of the time-dependent variable, usually hours}

\item{cvar}{Continuous variable of interest (x ~ y)}

\item{group}{Grouping variable to apply to the `cvar` (x ~ y)}
}
\description{
`circadian_ttest` takes data and returns a table of t-test based on the continuous variable of interest and a grouping variable.
}
\details{
Applies a simple data transformation to identify the summary statistics of the data frame by the stated variables. Results in a mean, standard deviation, and standard error term. This data is also used for making a t-test based table, which can then also be graphed in [autonomic::geom_circadian].
}
