% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{compute_Biais}
\alias{compute_Biais}
\title{Biais}
\usage{
compute_Biais(obs, sim, na.rm = TRUE, sim_minus_obs = TRUE)
}
\arguments{
\item{obs}{Observed streamflow vector}

\item{sim}{Simulated streamflow vector}

\item{na.rm}{Should missing values be omited ?}

\item{sim_minus_obs:}{should it be sim - obs ? (or the other way
around)}
}
\description{
Computes the bias (unitless) between simulated and
observed data
}
