% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elasticity.R
\name{compute_elasticity}
\alias{compute_elasticity}
\title{Elasticity}
\usage{
compute_elasticity(Q, X)
}
\arguments{
\item{Q}{discharge}

\item{X}{variable}
}
\value{
Elasticity given by :
eps_X = median((Q-Qmean)/(X-Xmean) * Xmean/Qmean)
}
\description{
To quote Yushiou Tsai (2016) :
"We used elasticity to assess the sensitivity of mean and drought
flows to changes in climate, land use and land cover (LULC), and
water use across a broad region. \link{...} Sankarasubramanian et al.
(2001) suggested (that) a bivariate empirical estimator for the
precipitation elasticity of mean flow".
This is this last estimator that is use in the below function.
}
