% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_duration_curve.R
\name{compute_FDC}
\alias{compute_FDC}
\title{fdc_values}
\usage{
compute_FDC(Q, n = 1000, sort = FALSE, isNormLaw = FALSE, na.rm = TRUE)
}
\arguments{
\item{Q}{Streamflow vector}

\item{n}{number of rows in the resulting data.frame (should be
smaller than the length of 'Q'.}

\item{sort}{logical. Should the sort function be used instead
of the quantile function ?}

\item{na.rm}{logical. Should the missing values be ignored ? (must
be TRUE if the quantile function is used !)}
}
\value{
res
}
\description{
Given a vector of streamflow values, computes a
data.frame with two columns : a 'p' column containing the
probability of exceedance and a 'Q' column containing the
corresponding streamflow values. Two methods can be used : simply
sorting the data (not recommended) or using the quantile function.
}
