% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_duration_curve.R
\name{compute_FDC_p}
\alias{compute_FDC_p}
\title{compute_FDC_p}
\usage{
compute_FDC_p(n = 1000, sort = FALSE, isNormLaw = FALSE, na.rm = TRUE)
}
\arguments{
\item{Q}{discharge}
}
\description{
description
}
\seealso{
ref
}
