% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_period.R
\name{compute_GumbelLaw}
\alias{compute_GumbelLaw}
\title{Compute Quantile from Gumbel Law}
\usage{
compute_GumbelLaw(a, b, returnPeriod)
}
\arguments{
\item{a}{Numeric value of the Gumbel distribution location parameter.}

\item{b}{Numeric value of the Gumbel distribution scale parameter.}

\item{returnPeriod}{Numeric value of the return period (in years or consistent time unit).}
}
\value{
Numeric value representing the estimated discharge for the given return period.
}
\description{
Calculates the discharge value (quantile) associated with a given return period using the Gumbel distribution parameters.
}
