% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_period.R
\name{compute_GumbelParams}
\alias{compute_GumbelParams}
\title{Estimate Gumbel Distribution Parameters}
\usage{
compute_GumbelParams(X)
}
\arguments{
\item{X}{Numeric vector representing the maximum flows (e.g., annual daily maxima).}
}
\value{
A \link{list} with \code{a} the location parameter and \code{b} the scale parameter).
}
\description{
Computes the location (\code{a}) and scale (\code{b}) parameters of the Gumbel distribution based on a numeric vector of annual maximum flows.
}
