% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{compute_hsaLog}
\alias{compute_hsaLog}
\title{hsaLog}
\usage{
compute_hsaLog(x, method = "inf.na")
}
\arguments{
\item{x}{Any streamflow vector}

\item{method}{the method to use : 'Pushpalatha2012', a numeric
value (to be added to 'x') or 'inf.na'}
}
\value{
log(x) according to parametrization
}
\description{
Computes the log transformed value of streamflow using
either the approach proposed by Pushpalatha et al. (2012), or by
adding any given value to all streamflow values or simply by
setting to NA any non finite value resulting from the application
of log().
}
