% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{compute_KGE}
\alias{compute_KGE}
\title{KGE}
\usage{
compute_KGE(obs, sim, na.rm = TRUE, method = 1)
}
\arguments{
\item{obs}{Observed streamflow vector}

\item{sim}{Simulated streamflow vector}

\item{na.rm}{Should missing values be omited?}

\item{method}{Two methods are implemented (see hydroGOF::KGE): '1'
for Gupta et al. (2009) and '2' for Kling et al. (2012)}
}
\value{
KGE
}
\description{
Computes the Kling-Gupta efficiency coefficient. This
function was largely inspired by the similar function in the
HydroGOF package. Here, only a simplified version is provided with
no checks on inputs and very little formatting to speed it up.
}
