% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{compute_KGEracine}
\alias{compute_KGEracine}
\title{KGEracine}
\usage{
compute_KGEracine(obs, sim, na.rm = TRUE)
}
\arguments{
\item{obs}{Observed streamflow vector}

\item{sim}{Simulated streamflow vector}

\item{na.rm}{Should missing values be omited ?}
}
\value{
KGEracine according to parametrization
}
\description{
Computes the
}
