% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_period.R
\name{compute_LogNormal}
\alias{compute_LogNormal}
\title{Calculation of a value according to a return period according to a Galton distribution (log-normal)}
\usage{
compute_LogNormal(X, returnPeriod)
}
\arguments{
\item{X}{\link{numeric} vector of annual data to process}

\item{returnPeriod}{\link{numeric} Return period in years}
}
\value{
The value for the return period according to Galton distribution
}
\description{
Calculation of a value according to a return period according to a Galton distribution (log-normal)
}
