% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recession.R
\name{compute_median_dtRec}
\alias{compute_median_dtRec}
\title{compute_median_dtRec}
\usage{
compute_median_dtRec(Q, na.rm = TRUE)
}
\arguments{
\item{Q}{discharge}
}
\description{
description
}
\seealso{
ref
}
