% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{compute_NSEi}
\alias{compute_NSEi}
\title{NSEi}
\usage{
compute_NSEi(obs, sim, na.rm = TRUE)
}
\arguments{
\item{obs}{Observed streamflow vector}

\item{sim}{Simulated streamflow vector}

\item{na.rm}{Should missing values be omited ?}
}
\value{
NSEi according to parametrization
}
\description{
Computes the Nash-Sutcliffe efficiency coefficient on
}
