% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{compute_NSElog}
\alias{compute_NSElog}
\title{NSElog}
\usage{
compute_NSElog(obs, sim, na.rm = TRUE, log_method = "inf.na")
}
\arguments{
\item{obs}{Observed streamflow vector}

\item{sim}{Simulated streamflow vector}

\item{na.rm}{Should missing values be omited ?}

\item{log_method}{See 'hsaLog'}
}
\value{
NSElog according to parametrization
}
\description{
Computes the Nash-Sutcliffe efficiency coefficient on
the log transformed streamflow values using the 'hsaLog' function
and the 'NSE' function
}
