% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{compute_NSEracine}
\alias{compute_NSEracine}
\title{NSEracine}
\usage{
compute_NSEracine(obs, sim, na.rm = TRUE)
}
\arguments{
\item{obs}{Observed streamflow vector}

\item{sim}{Simulated streamflow vector}

\item{na.rm}{Should missing values be omited ?}
}
\value{
NSEracine according to parametrization
}
\description{
Computes the Nash-Sutcliffe efficiency coefficient on
}
