% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent_survival_table.R
\docType{data}
\name{data}
\alias{data}
\title{Create data table of recurrent events for survival analysis}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 313 rows and 10 columns.
}
\usage{
data
}
\arguments{
\item{data}{A dataframe containing the subsequent parameters}

\item{id}{Column in dataframe that contains unique IDs for each row}

\item{first}{Column with left/enrollment dates}

\item{last}{Column with right/censoring time point, or last contact}

\item{event.dates}{Vector of columns that contain event dates}

\item{model.type}{Character/string = c("marginal", "pwptt", "pwpgt")}

\item{death}{Column created if death is known (0 or 1), Must include in original dataframe (e.g. can add column of zeroes prn)}
}
\value{
A data frame organized into a survival table format
}
\description{
Create data table of recurrent events for survival analysis
}
\examples{
recurrent_survival_table(tibble, id, first.date, last.known.date, c(event.dates), "marginal", death.status)
}
\keyword{datasets}
