% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataDemo}
\alias{dataDemo}
\title{Hourly time series data with clinical covariates}
\format{
An tibble data frame
}
\usage{
dataDemo
}
\description{
Data is from an algorithm that generates a summary HRV measure using the Poincare phase-space plot, generated from kurtoses of the x and y axis. Clinical data is also available for visualization and comparison. There are repeat rows for each hour that Dyx was taken.
}
\keyword{datasets}
