% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.R
\name{.empty_data_frame}
\alias{.empty_data_frame}
\title{Build Empty Data Frame with Named Columns (Internal)}
\usage{
.empty_data_frame(col_names)
}
\arguments{
\item{col_names}{Character vector of column names.}
}
\value{
Empty data.frame.
}
\description{
Build Empty Data Frame with Named Columns (Internal)
}
\keyword{internal}
