% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.R
\name{.extract_column_definitions}
\alias{.extract_column_definitions}
\title{Extract Column Definitions from Tab Format}
\usage{
.extract_column_definitions(metadata, column_line)
}
\arguments{
\item{metadata}{Character vector of metadata lines}

\item{column_line}{The column header line}
}
\value{
Named character vector of column descriptions
}
\description{
Extract Column Definitions from Tab Format
}
\keyword{internal}
