% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.R
\name{.extract_meta_info}
\alias{.extract_meta_info}
\title{Extract Meta Information from Header}
\usage{
.extract_meta_info(metadata)
}
\arguments{
\item{metadata}{Character vector of metadata lines}
}
\value{
List of meta information
}
\description{
Extract Meta Information from Header
}
\keyword{internal}
