% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.R
\name{.open_vep_connection}
\alias{.open_vep_connection}
\title{Open VEP File Connection (Internal)}
\usage{
.open_vep_connection(file)
}
\arguments{
\item{file}{File path.}
}
\value{
Open connection.
}
\description{
Open VEP File Connection (Internal)
}
\keyword{internal}
