% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.R
\name{.parse_tab_header}
\alias{.parse_tab_header}
\title{Parse Tab-Delimited VEP Header}
\usage{
.parse_tab_header(metadata, column_line)
}
\arguments{
\item{metadata}{Character vector of metadata lines}

\item{column_line}{The column header line}
}
\value{
Parsed header list
}
\description{
Parse Tab-Delimited VEP Header
}
\keyword{internal}
