% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.R
\name{.parse_tabular_data}
\alias{.parse_tabular_data}
\title{Parse Tabular Data Block (Internal)}
\usage{
.parse_tabular_data(data_lines, col_names)
}
\arguments{
\item{data_lines}{Character vector of data lines.}

\item{col_names}{Character vector of output column names.}
}
\value{
Parsed data.frame.
}
\description{
Parse Tabular Data Block (Internal)
}
\keyword{internal}
