% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.R
\name{.parse_vcf_header}
\alias{.parse_vcf_header}
\title{Parse VCF Format VEP Header}
\usage{
.parse_vcf_header(metadata, column_line)
}
\arguments{
\item{metadata}{Character vector of metadata lines}

\item{column_line}{The column header line}
}
\value{
Parsed header list
}
\description{
Parse VCF Format VEP Header
}
\keyword{internal}
