% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.R
\name{.read_tab_data}
\alias{.read_tab_data}
\title{Read Tab-Delimited VEP Data (Internal)}
\usage{
.read_tab_data(con, header, parse_extra)
}
\arguments{
\item{con}{Connection to read from (already positioned at data start)}

\item{header}{Header structure from read_vep_header}

\item{parse_extra}{Whether to parse Extra column}
}
\value{
Data frame
}
\description{
Read Tab-Delimited VEP Data (Internal)
}
\keyword{internal}
