% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_duration_curve.R
\name{fdc_slope}
\alias{fdc_slope}
\title{fdc_slope}
\usage{
fdc_slope(Q, p = c(0.33, 0.66))
}
\arguments{
\item{Q}{Streamflow vector}

\item{p}{A length 2 numeric vector containing the exceedance
probability that define the of the mid-segment}
}
\value{
Mid-segment low duration curve slope
}
\description{
Compute the mid-segment flow duration curve slope
}
