% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-marksman.R
\name{fire}
\alias{fire}
\title{Firing using the Aims at Target Hypotheses}
\usage{
fire(aim, data, ...)
}
\arguments{
\item{aim}{A single \code{aim} object}

\item{data}{Data frame that contains the variables that were planned for in
the specific aims generated prior.}

\item{...}{For extensibility}
}
\value{
Returns the fitted models in an expanded table, still of the \code{aim} class
}
\description{
As the analysis time can be quite long, this function separates
out model runtime until after the workflows for analysis are completed.
This function works on a single \code{aim} object at a time, allowing for list
methods such as \link[purrr:map]{purrr::map} to apply.
}
