% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee.R
\name{gee_reg}
\alias{gee_reg}
\alias{update.gee_reg}
\alias{print.gee_reg}
\title{General Interface for Generalized Estimating Equations}
\usage{
gee_reg(mode = "regression", cluster = NULL, correlation = NULL)

\method{update}{gee_reg}(object, cluster = NULL, correlation = NULL, fresh = FALSE, ...)

\method{print}{gee_reg}(x, ...)
}
\arguments{
\item{mode}{A character string that describes the type of model. In this case,
it currently supports type of "regression", which is default.}

\item{cluster}{Name of the column that identifies the clustering variable. The
column in the dataset that corresponds to the clustering variable should be
of equal length as the other observations.}

\item{correlation}{A character string specifying the correlation structure.
Currently supports the following:
\itemize{
\item \code{gee}: "independence", "fixed", "stat_M_dep", "non_stat_M_dep", "exchangeable", "AR-M", "unstructured"
\item \code{geepack}: "independence", "exchangeable", "ar1", "unstructured", "userdefined"
}}

\item{object}{Generalized estimating equation model specification}

\item{fresh}{A logical for whether the arguments should be modified in place or replaced altogether}

\item{...}{Extensible}

\item{x}{GEE model specification}
}
\description{
\code{gee_reg()} is a \emph{parsnip} friendly method for specification of
generalized estimating equation model before fitting. It currently uses the \code{gee} and \code{geepack} packages for its engines.
}
\examples{
library(parsnip)
library(magrittr)
data(geh)
f <- qrs_tang ~ lab_hba1c
gee_mod <-
  gee_reg(cluster = hhp_id, correlation = "independence") \%>\%
  set_engine("geepack")

}
