% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian_plots.R
\name{geom_circadian}
\alias{geom_circadian}
\title{Circadian Plot by Group}
\usage{
geom_circadian(time = "timegrp", group = "catvar")
}
\value{
Returns a ggplot object of geom type, other layers can be added on as seen in example.
}
\description{
`geom_circadian` converts the output of [autonomicR::circadian_ttest] into a complex geom that is broken down by time/hour and HRV (or any other continuous variable). Each hour is then separated by the grouping variable.
}
\details{
Currently creates a geom_layer that shows a error bar and point estimate of values by group (e.g. clinical status). As t-test values should be available in the data frame, shows points of significance.
}
\examples{
ggplot(df_ttest) + geom_circadian()
}
