% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circadian_plots.R
\name{geom_forest}
\alias{geom_forest}
\title{Forest Plot of Hourly Odds}
\usage{
geom_forest(group = "time", or = "OR", lower = "Lower", upper = "Upper", ...)
}
\arguments{
\item{group}{Is the grouping variable from data in ggplot, like hour}

\item{or}{Odds ratio}

\item{lower}{Lower boundary of 95 percent CI}

\item{upper}{Upper boundary of 95 percent CI}
}
\value{
A ggplot geom that has been configured for a forest plot.
}
\description{
\code{geom_forest} creates an OR plot for each hour of data given. Its a
ggplot format so additional variables, like titles, can be added in.
}
\details{
This function creates a forest plot using the OR developed
by the \code{circadian_odds} function in this package. By default, it
takes the output of \verb{circadian odds}, which is a tibble named "ot",
and will generate a forest plot based on the grouping variable
(default is time of day). Original data can be restricted or the
hours can be reduced).
}
\examples{
ggplot(OddsTable) + geom_forest("hour", "OddsRatio", "LowerCI", "UpperCI")
}
