% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{get_BFI}
\alias{get_BFI}
\title{baseflow_index}
\usage{
get_BFI(Q, BF, na.rm = TRUE)
}
\arguments{
\item{Q}{Streamflow vector}

\item{BF}{Baseflow vector}

\item{na.rm}{Should missing values be omited ?}
}
\value{
res
}
\description{
Compute the baseflow index: the ratio between the
total baseflow volume and the total streamflow volume.
Warning : No checks are done on inputs ! You should have selected
the proper period before hand (i.e. both vector should be of same
length and span over the same period).
}
