% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{get_deltaX}
\alias{get_deltaX}
\title{get_deltaX}
\usage{
get_deltaX(
  X,
  Date,
  past,
  futur,
  to_normalise,
  returnPeriod = NULL,
  waterType = "low",
  Q_for_BFI = NULL
)
}
\arguments{
\item{Q}{discharge}
}
\description{
description
}
\seealso{
ref
}
