% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD.SCMapping.R
\name{getHighresCords}
\alias{getHighresCords}
\title{The function to sample the spatial location information for each single cell}
\usage{
getHighresCords(Cords, numCell, shape = "Square")
}
\arguments{
\item{Cords}{The spatial location information in the measure spatial locations, with the first and second columns represent the 2-D x-y coordinate system}

\item{numCell}{a numeric value indicating the number of single cells in each measured location, we suggest 20 for ST technology, 7 for 10x Viisum and 2 for Slide-seq}

\item{shape}{a character indicating whether the sampled spatial coordinates for single cells locating in a Square-like region or a Circle-like region. The center of this region is the measured spatial location in the non-single cell resolution spatial transcriptomics data. The default is "Square", the other shape is "Circle"}
}
\value{
Returns a dataframe with the sampled spatial location information for each single cell
}
\description{
The function to sample the spatial location information for each single cell
}
