% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-diagnostics.R
\name{ggcosinorfit}
\alias{ggcosinorfit}
\title{Graphical Assessment of Amplitude and Acrophase}
\usage{
ggcosinorfit(model)
}
\arguments{
\item{model}{Requires a cosinor model to extract the correct statistics to
generate the plot.}
}
\value{
ggplot object that has has multiple components
}
\description{
Graphical Assessment of Amplitude and Acrophase
}
\examples{
data("twins")
m <- cosinor(rDYX ~ hour, twins)
ggcosinorfit(model)

}
