% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-diagnostics.R
\name{ggmulticosinor}
\alias{ggmulticosinor}
\title{ggplot of multiple cosinor models}
\usage{
ggmulticosinor(objects, ...)
}
\arguments{
\item{objects}{List of \code{cosinor} objects}

\item{...}{For extensibility}
}
\value{
Object of class \code{ggplot} that can be layered
}
\description{
ggplot of multiple cosinor objects for comparison purposes. Appearance of plot will be greatly enhanced if objects share a similar y-axis scale.
}
