% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-diagnostics.R
\name{ggpopcosinor}
\alias{ggpopcosinor}
\title{ggplot of population cosinor model}
\usage{
ggpopcosinor(object, ...)
}
\arguments{
\item{object}{A \code{cosinor} object of the population type, called when the population is specified in \code{\link[card:cosinor]{card::cosinor()}}}

\item{...}{For extensibility}
}
\value{
Returns object of class \code{ggplot} that can be layered
}
\description{
ggplot of a population cosinor model
}
