% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-heart.R
\docType{class}
\name{heart-class}
\alias{heart-class}
\title{\code{heart} class}
\description{
\code{heart} is an S4 class that serves as an object that holds together different components that represent the anatomical heart. A \code{heart} object allows for storing clinical information about a patient together, which may more readily allow for analytical approaches to be developed. The heart has plumbing (the coronary arteries), electricity (the nerves), and walls (cardiac chambers). The slots represent these objects.
}
\section{Slots}{

\describe{
\item{\code{pipes}}{the epicardial and resistance vessels}

\item{\code{electric}}{the conduction system of nerves}

\item{\code{structural}}{the cardiac chambers and valves}
}}

