% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hrv}
\alias{hrv}
\title{Output from MATLAB HRV Toolbox}
\format{
An tibble data frame
}
\usage{
hrv
}
\description{
Data is a single patient data output from HRV Toolbox. It contains granular data of calculated HRV in 5-second sliding windows.
}
\keyword{datasets}
