% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrv_modeling.R
\name{hrv_model_building}
\alias{hrv_model_building}
\title{HRV Model Building}
\usage{
hrv_model_building(data, covar.builds, model, prop.scores = NULL)
}
\description{
\code{hrv_model_building} creates a list of models using different builds of covariates.
}
