% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maude_complication_index}
\alias{maude_complication_index}
\title{MAUDE Complication Index for Cardiovascular Procedure Adverse Events}
\format{
An object of class \code{list} of length 16.
}
\usage{
maude_complication_index
}
\description{
A named list of normalized MAUDE problem terms organized into the
complication categories defined in \code{complication_definitions}, plus a
residual \code{not_indexed} bucket. The list is maintained explicitly in
\code{data-raw/complications.R} as hand-written term vectors. Overlap between
categories is expected.
}
\examples{
maude_complication_index$pericardial
maude_complication_index$stroke
maude_complication_index$device_malfunction
maude_complication_index$not_indexed

}
\keyword{datasets}
