% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_hrv.R
\name{model_hrv}
\alias{model_hrv}
\title{Modeling HRV Variables}
\usage{
model_hrv(data, outcome, exposure, covar = NULL)
}
\arguments{
\item{data}{Data containing names of outcome, exposure, and covars}

\item{outcome}{Outcome of interest}

\item{exposure}{Exposure(s) of interest. If multiple given, will create a list of models to return.}

\item{covar}{Covariates to include in model}
}
\value{
List of models
}
\description{
Different HRV variables (e.g. HF, LF, SDNN) can be modeling as predictors for an outcome with adjustment by covariates. Can choose between logistic and linear based on the outcome variable as needed. Shortens process for modeling
}
\examples{
#models <- model_hrv(df, "death", c("HF", "LF"), c("hptn", "dm"))
}
