% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maude-adjudicate.R
\name{normalize_complication_classification}
\alias{normalize_complication_classification}
\title{Normalize complication classifications to a named character vector}
\usage{
normalize_complication_classification(classification)
}
\description{
Normalize complication classifications to a named character vector
}
\keyword{internal}
