% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maude-adjudicate.R
\name{normalize_maude_terms}
\alias{normalize_maude_terms}
\title{Utility function to help manage problem terms such they are normalized and spacing or common typos, capitalizations, etc., don't become an issue}
\usage{
normalize_maude_terms(terms)
}
\description{
Utility function to help manage problem terms such they are normalized and spacing or common typos, capitalizations, etc., don't become an issue
}
\keyword{internal}
