% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARD.imputation.R
\name{normCoordsTrainTest}
\alias{normCoordsTrainTest}
\title{Normalize the new spatial locations without changing the shape and relative positions}
\usage{
normCoordsTrainTest(locationOrig, trainInd, testInd)
}
\arguments{
\item{locationOrig}{Data frame, spatial location data frame of the original spatial resolved transcriptomics dataset, stored in the CARD_object@spatial_location}

\item{trainInd}{Vector, Index of the original spatial locations}

\item{testInd}{Vector, Index of the newly grided spatial locations}
}
\value{
Return the normalized spatial location data frame
}
\description{
Normalize the new spatial locations without changing the shape and relative positions
}
