% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecg_patches.R
\name{patch_read_vivalnk}
\alias{patch_read_vivalnk}
\title{Read in VivaLNK Patch Data}
\usage{
patch_read_vivalnk(name, loc)
}
\arguments{
\item{name}{Name of the patient that the log file was created on. The variable does NOT have an extension.}

\item{loc}{The raw, unprocessed data folder containining the log file (which is in a ".txt" format)}
}
\value{
Returns a simple data frame wiith the start and endtime of the ECG data. Has several additional markers available internally, but currently not called.
}
\description{
`patch_read_vivalnk` Reads in VivaLNK log files, raw, and helps to organize the summary data.
}
\details{
Does not obtain the ECG signal, just the summary data from the log file for diagnostic purposes.
}
