% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab-toolbox.R
\name{proc_hrv_matlab}
\alias{proc_hrv_matlab}
\title{Process Toolbox HRV}
\usage{
proc_hrv_matlab(loc, name, time = 3600)
}
\arguments{
\item{loc}{Location of the folder that contains all of the patients
that were analyzed by the Main_HRV_Analysis.m function from the
Toolbox.}

\item{name}{Name of the patient/ID. There should exist a folder with
the name inside the \code{loc} folder. Inside this folder are all the
Toolbox parameters and HRV results in CSV format.}

\item{time}{Number of seconds to group the HRV data by. Defaults to
3600 seconds (which is 1 hour)}
}
\value{
Data frame of HRV summarized by the grouping variable (e.g.
3600 seconds = 1 hour). Also returns an additional column of
percent missing (e.g. 20.0\% missing data) by time group.
}
\description{
Takes the output from HRV Toolbox and converts it for analysis.
Uses the package \code{\link[data.table:fread]{data.table::fread()}} for reading in data due to
size/speed.
}
\details{
The data is taken sequentially (sliding windows), and
summarized over the course of certain time lengths. The data comes
in a standardized pattern from the toolbox. It requires processing
due to its large file sizes (e.g. 24 hours of data for a single
patient can be up to 2 MB in size).
}
\examples{
# Setup
# library(data.table)
# data(hrv)

# Assuming in "root/code" folder, need to get to data in STUDY
parent <- dirname(getwd())
loc <- file.path(parent, "data", "proc_hrv", "STUDY")
vars <- Hmisc::Cs(SDNN, RMSSD, pnn50, ulf, vlf, lf, hf, lfhf, ttlpwr, ac, dc, SampEn, ApEn)
time <- 3600 # number of seconds in an hour
names <- list.files(path = loc)

# Create for loop to eventually make data frame (not currently working)
# d <- list()
# for(i in 1:length(names)) {
#  d[[paste(names[i])]] <- proc_hrv_matlab(loc, names[i], vars, time)
# }
# df <- rbindlist(d) \%>\% as_tibble()
}
