% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent-events.R
\name{recur_summary}
\alias{recur_summary}
\title{Recurrent Event Summary Table by Group}
\usage{
recur_summary(data, covar)
}
\arguments{
\item{data}{Recurrent event data in marginal format. There must be
an ID column. Must merge in the covariate of interest into this data set.}

\item{covar}{Name of covariate of interest to serve as grouping variable.}
}
\value{
Summary table by grouping variable, can be placed into a latex
environment with kableExtra. Assumes that death events may be
present when most recent non-EVENT has status 1.
}
\description{
\code{recur_summary} Creates a table with summary of
recurrent events
}
\details{
This function allows for taking the output of
\link[card:recur_survival_table]{card::recur_survival_table} marginal format repeat event data,
and creates a summary table that describes the number of events by
strata/event.
}
