% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent-events.R
\name{recurrent_propensity}
\alias{recurrent_propensity}
\title{Propensity Score Weighting}
\usage{
recurrent_propensity(data, vars)
}
\arguments{
\item{data}{Data frame that contains all covariates and outcomes. First column should be ID}

\item{vars}{Variables used for regression. Outcome variable must be first.}
}
\value{
Returns a modified table from what was originally given with the new columns propensity scores. Essentially original df + 2 columns.
}
\description{
\code{recurrent_propensity} Adds propensity score to any data set that is being regressed upon.
}
\details{
Using a logistic regression, will take covariates and create propensity scores, and adds the weights. Uses the standard logistic regression to evaluate the propensity score.
}
