% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent_survival_data.R
\name{recurrent_summary_table}
\alias{recurrent_summary_table}
\title{Recurrent Event Summary Table by Group}
\usage{
recurrent_summary_table(marginal.data, group)
}
\arguments{
\item{marginal.data}{Recurrent event data in marginal format. ID column must be present as first column.}

\item{group}{Table that has ID in first column, and named covar as second column.}
}
\value{
Summary table by grouping variable, can be placed into a latex environment with kable and kable styling. Assumes that death events may be present when most recent non-EVENT has status 1.
}
\description{
\code{recurrent_summary_table} Creates a table with summary of recurrent events
}
\details{
This function allows for taking the output of \code{recurrent_survival_table}, marginal format repeat event data, and creates a summary table that describes the number of events by strata/event.
}
\examples{
tbl <- recurrent_summary_table(marg, grp)
tbl \%>\% kable("latex", caption = "Title", booktabs = TRUE) \%>\%
kable_styling(font_size = 8)
}
