% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-marksman.R
\name{reload}
\alias{reload}
\alias{reload.default}
\alias{reload.bullet}
\alias{reload.aim}
\title{Reload either the bullets or the aims}
\usage{
reload(mark, ...)

\method{reload}{default}(mark, ...)

\method{reload}{bullet}(mark, ...)

\method{reload}{aim}(mark, ...)
}
\arguments{
\item{mark}{Represents either a \code{bullet} or \code{aim} object to be analyzed for
further information or characteristics
\itemize{
\item A single \strong{bullet} object.
\item A single \strong{aim} object. If it has been processed by \code{fire()}, then that
component will be removed. Remember, this exists as one of the named lists
created from the \code{aim()} function.
}}

\item{...}{Any original parameter from \code{aim} or \code{bullet} can be
added/revised. Both require calling the name of the element to be revised.
The function will assess the names and update accordingly. Only a single
element can be modified at a time.
\itemize{
\item If an \strong{aim} object: c("outcomes", "model_num", "predictors", "models",
"raw")
\item If a \strong{bullet} object: c("outcomes", "predictors", "exposures",
"covariates", "approach", "model"), e.g. \code{reload(bullet, outcomes =   "new_outcome")}
}}
}
\value{
Returns the original \code{aim} or \code{bullet} with revised components.
}
\description{
If, after an \code{aim} or \code{bullet} object has been made, it can be
revised or reloaded with new characteristics
}
\concept{reload}
