% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatEventsData.R
\docType{data}
\name{repeatEventsData}
\alias{repeatEventsData}
\title{Recurrent event sample data}
\format{
An tibble data frame
}
\usage{
data(repeatEventsData)
}
\description{
Data is from a outcomes study on cardiovascular outcomes.
It contains the first visit date, the last known date, and
times of various events that have happened. They document death at
right censoring as well. These events are non-ordered.
}
\examples{
data(repeatEventsData)
}
\keyword{datasets}
