% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/af-ctrs.R
\name{set_afib_score_parameters}
\alias{set_afib_score_parameters}
\title{Set the weights for the \code{afib_score()} function}
\usage{
set_afib_score_parameters(
  weight_symptoms = NULL,
  weight_afeqt_total = NULL,
  weight_afeqt_symptoms = NULL,
  weight_afeqt_activities = NULL,
  weight_afeqt_treatment = NULL,
  weight_nyha_class = NULL,
  weight_electrical = NULL,
  mct_cutoff_low = NULL,
  mct_cutoff_medium = NULL,
  ecg_cutoff_low = NULL,
  ecg_cutoff_medium = NULL
)
}
\arguments{
\item{weight_afeqt_total}{A \code{numeric} value representing the weight of the
total AFEQT score in the composite score.}

\item{weight_afeqt_symptoms}{A \code{numeric} value representing the weight of the
AFEQT symptoms domain in the composite score.}

\item{weight_afeqt_activities}{A \code{numeric} value representing the weight of
the AFEQT daily activities domain in the composite score.}

\item{weight_afeqt_treatment}{A \code{numeric} value representing the weight of
the AFEQT treatment concerns domain in the composite score.}

\item{weight_nyha_class}{A \code{numeric} value representing the weight of the
NYHA class in the composite score.}

\item{weight_electrical}{A \code{numeric} value representing the weight of the
electrical burden in the composite score.}

\item{mct_cutoff_low}{A \code{numeric} value representing the cutoff for low MCT
burden.}

\item{mct_cutoff_medium}{A \code{numeric} value representing the cutoff for medium
MCT burden.}

\item{ecg_cutoff_low}{A \code{numeric} value representing the cutoff for low ECG
burden.}

\item{ecg_cutoff_medium}{A \code{numeric} value representing the cutoff for medium
ECG burden.}
}
\description{
This function sets the global parameters for the \code{\link[=afib_score]{afib_score()}} function.
}
\examples{
set_afib_score_parameters(
  weight_symptoms = 0.5,
  weight_electrical = 0.5,
  weight_afeqt_total = 0.5, weight_afeqt_symptoms = 0.2,
  weight_afeqt_activities = 0.2, weight_afeqt_treatment = 0.1, weight_nyha_class = 0.7,
  weight_electrical = 0.5, mct_cutoff_low = 0.33, mct_cutoff_medium = 0.66,
  ecg_cutoff_low = 0.33, ecg_cutoff_medium = 0.66
)

}
