% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-circles.R
\name{tidy.circular}
\alias{tidy.circular}
\title{Tidy a(n) circular object}
\usage{
\method{tidy}{circular}(x, conf.int = FALSE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{A \code{circular} regression object}

\item{conf.int}{Logical indicating whether or not to include confidence
interval in tidied output}

\item{conf.level}{The confidence level to use if \code{conf.int = TRUE}. Must be
between 0 and 1, with default to 0.95 (the 95\% confidence interval).}

\item{...}{For extensibility}
}
\value{
a \code{tibble} object
}
\description{
Tidy summarizes information about the components of a \code{circular}
model.
}
\details{
\code{circular} object are made using the \code{\link[circular:lm.circular]{circular::lm.circular()}} function
}
