% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{triplets}
\alias{triplets}
\title{Hourly time series data with clinical covariates}
\format{
An tibble data frame
}
\usage{
triplets
}
\description{
Clinical data is also available for visualization and comparison. Other HRV measures are used here for comparison and testing out functions.
}
\keyword{datasets}
