% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_named_list}
\alias{validate_named_list}
\title{Validate that a list is a named list where each element has is named}
\usage{
validate_named_list(x)
}
\description{
Validate that a list is a named list where each element has is named
}
