% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zipcode}
\alias{zipcode}
\title{Zipcodes with Associated Latitude and Longitude}
\format{
A data frame with character vector zipcodes and latitude/longitude
}
\usage{
zipcode
}
\description{
This is a dataset from the archived/orphaned {zipcode} package.
}
\keyword{datasets}
