% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maude.R
\name{adjudicate_maude_event}
\alias{adjudicate_maude_event}
\title{Adjudicate MAUDE adverse event narratives using a large language model (LLM)}
\usage{
adjudicate_maude_event()
}
\description{
Adjudicate MAUDE adverse event narratives using a large language model (LLM)
}
