% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{apply_threshold}
\alias{apply_threshold}
\title{apply_threshold}
\usage{
apply_threshold(
  X,
  lim,
  where = "<=",
  what = "X",
  select = "all",
  Date = NULL,
  period = NULL
)
}
\arguments{
\item{Q}{discharge}
}
\description{
description
}
\seealso{
ref
}
