% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-package.R
\docType{package}
\name{card-package}
\alias{card}
\alias{card-package}
\title{card: Cardiovascular Research Data}
\description{
A collection of cardiovascular research datasets and analytical tools, including methods for cardiovascular procedural data, such as electrocardiography, echocardiography, and catheterization data. Additional methods exist for analysis of procedural billing codes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cran.r-project.org/package=card}
  \item Report bugs at \url{https://github.com/asshah4/card/issues}
}

}
\author{
\strong{Maintainer}: Anish S. Shah \email{ashah282@uic.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID}) [copyright holder]

}
\keyword{internal}
