% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autonomicR-package.R
\docType{package}
\name{card-package}
\alias{card}
\alias{card-package}
\title{card: Cardiovascular and Autonomic Research Designs}
\description{
This package contains tools needed for the management of
	data that has been collected to study autonomic physiology, 
	particularly of how the autnomic nervous system effects the 
	cardiovascular system (neurocardiology). The package uses tools
	built to study electrocardiographic time series data, focused on
	heart rate variability and morphology in companionship with MATLAB.
	It also helps with the management of extended time series, such as
	longitudinal data and recurrent events. It also contains methods
	for analyzing and studying circadian events.
}
\keyword{internal}
