% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-package.R
\docType{package}
\name{card-package}
\alias{card}
\alias{card-package}
\title{card: Cardiovascular and Autonomic Research Design}
\description{
Tools that can aid in the assessment of the autonomic regulation 
		of cardiovascular physiology. The aims of this package are to: 1) study 
		ECG-based time signals (both RR intervals and morphology) as extensions 
		of signal processing, 2) study circadian rhythms and how it effects 
		autonomic physiology, 3) assess clinical risk of autonomic dysfunction on 
		cardiovascular health through the perspective of epidemiology and causality.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asshah4/card}
  \item Report bugs at \url{https://github.com/asshah4/card/issues}
}

}
\author{
\strong{Maintainer}: Anish S. Shah \email{asshah4@emory.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID})

}
\keyword{internal}
