% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card-package.R
\docType{package}
\name{card-package}
\alias{card}
\alias{card-package}
\title{card: Circadian and Autonomic Research Design}
\description{
Tools that can aid in the assessment of the autonomic regulation of cardiovascular physiology. The aims of this package are to: 1) study electrocardiography (both intervals and morphology) as extensions of signal processing, 2) study circadian rhythms and how it effects autonomic physiology, 3) assess clinical risk of autonomic dysfunction on cardiovascular health through the perspective of epidemiology and causality. The analysis of circadian rhythms through cosinor analysis are built upon the methods by Cornelissen (2014) \doi{10.1186/1742-4682-11-16} and Refinetti, Cornelissen, Halberg (2014) \doi{10.1080/09291010600903692}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cran.r-project.org/package=card}
  \item Report bugs at \url{https://github.com/asshah4/card/issues}
}

}
\author{
\strong{Maintainer}: Anish S. Shah \email{asshah4@emory.edu} (\href{https://orcid.org/0000-0002-9729-1558}{ORCID}) [copyright holder]

}
\keyword{internal}
