% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{CARD.visualize.Cor}
\alias{CARD.visualize.Cor}
\title{Visualize the cell type proportion correlation}
\usage{
CARD.visualize.Cor(proportion, colors = colors)
}
\arguments{
\item{proportion}{Data frame, cell type proportion estimated by CARD in either original resolution or enhanced resolution.}

\item{colors}{Vector of color names that you want to use, if NULL, we will use the default color scale c("#91a28c","white","#8f2c37")}
}
\value{
Returns a ggcorrplot figure.
}
\description{
Visualize the cell type proportion correlation
}
