% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeitgeiber.R
\name{circ_center}
\alias{circ_center}
\title{Center Time Around a Zeitgeiber}
\usage{
circ_center()
}
\description{
Based on a centering time point, shifts a vector to a "before" and "after" system to help align multiple individuals to a universal time, like the sunrise or any other appropriate \emph{zeitgeiber}.
}
