% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_period.R
\name{compute_GumbelLaw}
\alias{compute_GumbelLaw}
\title{compute_GumbelLaw}
\usage{
compute_GumbelLaw(a, b, returnPeriod)
}
\arguments{
\item{a}{The location parameter of Gumbel law}

\item{b}{The scale parameter of Gumbel law}

\item{returnPeriod}{\link{numeric} Return period. Its unit is given by
the aggregation time step used for the Gumbel law.}
}
\value{
Gumbel law value
}
\description{
Computes Gumbel law
}
