% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_period.R
\name{compute_GumbelParams}
\alias{compute_GumbelParams}
\title{compute_GumbelParams}
\usage{
compute_GumbelParams(X)
}
\arguments{
\item{X}{Maximum flow at a given time step (usually the maximum
daily flow per year)}
}
\value{
A \link{list} with \code{a} the location parameter and \code{b} the scale
parameter
}
\description{
Computes Gumbel's laws parameters
}
\examples{
# load data
data(L0123001, package="airGR")
# define vector of flows, time step is the day
flows = BasinObs$Qls
# define vector of dates
dates= BasinObs$DatesR
QJXA = calcQJXA(x=dates, flows=flows)
gumbel = getGumbelParams(QJXA)
}
