% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_period.R
\name{compute_LogNormal}
\alias{compute_LogNormal}
\title{Estimate Return Value using Log-Normal Distribution}
\usage{
compute_LogNormal(X, returnPeriod)
}
\arguments{
\item{X}{Numeric vector of annual data (e.g., minimum or maximum discharge).}

\item{returnPeriod}{Numeric value of the return period in years.}
}
\value{
Numeric value for the return period according to Galton distribution.
}
\description{
Calculation of a value according to a return period according to a Galton distribution (log-normal)
}
