% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_dynamic.R
\name{compute_Rc}
\alias{compute_Rc}
\title{runoff_coefficient}
\usage{
compute_Rc(Q, R, na.rm = TRUE)
}
\arguments{
\item{Q}{Streamflow vector}

\item{R}{Precipitation vector}

\item{na.rm}{Should missing values be omited ?}
}
\value{
Runoff coefficient
}
\description{
Computes the runoff coefficient or ratio: the ratio
between the total strealfow volume and the total precipitaion
volume. Warning: No checks are done on inputs ! You should have
selected the proper period beforehand: both vector should be of
same length and span over the same period. In addition the runoff
coefficient should be compute from the start of a hydrological
year to the end of a hydrological year.
}
