% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{compute_tSnowmelt}
\alias{compute_tSnowmelt}
\title{compute_tSnowmelt}
\usage{
compute_tSnowmelt(
  X,
  p1,
  p2,
  d = 5,
  w = 0.9,
  a = 0.925,
  passes = 3,
  method = "Wal"
)
}
\arguments{
\item{Q}{discharge}
}
\description{
description
}
\seealso{
ref
}
