% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{compute_tVolSnowmelt}
\alias{compute_tVolSnowmelt}
\title{compute_tVolSnowmelt}
\usage{
compute_tVolSnowmelt(
  X,
  p,
  d = 5,
  w = 0.9,
  a = 0.925,
  passes = 3,
  method = "Wal"
)
}
\arguments{
\item{Q}{discharge}
}
\description{
description
}
\seealso{
ref
}
