% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{compute_VolSnowmelt}
\alias{compute_VolSnowmelt}
\title{compute_VolSnowmelt}
\usage{
compute_VolSnowmelt(X, d = 5, w = 0.9, a = 0.925, passes = 3, method = "Wal")
}
\arguments{
\item{Q}{discharge}
}
\description{
description
}
\seealso{
ref
}
