% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-construction.R
\name{cosinor}
\alias{cosinor}
\alias{cosinor.default}
\alias{cosinor.data.frame}
\alias{cosinor.matrix}
\alias{cosinor.formula}
\alias{cosinor.recipe}
\title{Fit a \code{cosinor}}
\usage{
cosinor(x, ...)

\method{cosinor}{default}(x, ...)

\method{cosinor}{data.frame}(x, y, ...)

\method{cosinor}{matrix}(x, y, ...)

\method{cosinor}{formula}(formula, data, ...)

\method{cosinor}{recipe}(x, data, ...)
}
\arguments{
\item{x}{Depending on the context:
\itemize{
\item A \strong{data frame} of time-based predictors/indices.
\item A \strong{matrix} of time-based predictors/indices.
\item A \strong{recipe} specifying a set of preprocessing steps
created from \code{\link[recipes:recipe]{recipes::recipe()}}.
}}

\item{...}{Not currently used, but required for extensibility.}

\item{y}{When \code{x} is a \strong{data frame} or \strong{matrix}, \code{y} is the outcome
specified as:
\itemize{
\item A \strong{data frame} with 1 numeric column.
\item A \strong{matrix} with 1 numeric column.
\item A numeric \strong{vector}.
}}

\item{formula}{A formula specifying the outcome terms on the left-hand side,
and the predictor terms on the right-hand side.}

\item{data}{When a \strong{recipe} or \strong{formula} is used, \code{data} is specified as:
\itemize{
\item A \strong{data frame} containing both the predictors and the outcome.
}}
}
\value{
A \code{cosinor} object.
}
\description{
\code{cosinor()} fits a regression model of a time variable to a continuous outcome use trigonometric features.
}
\examples{

# Data setup
data("twins")

# XY interface
# mod1 <- cosinor(twins["hour"], twins$rDYX)

# Formula interface
mod2 <- cosinor(rDYX ~ hour, twins)

# Recipes interface (still needs work)
# rec <- recipes::recipe(rDYX ~ hour, twins)
# rec <- recipes::step_log(rec, rDYX)
# mod3 <- cosinor(rec, twins)

}
