% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-diagnostics.R
\name{cosinor_area}
\alias{cosinor_area}
\title{Area of Ellipse}
\usage{
cosinor_area(object, level = 0.95, ...)
}
\arguments{
\item{object}{Model of class \code{cosinor}}

\item{level}{Confidence level requested}

\item{...}{Not currently used, but required for extensibility.}
}
\value{
Area of potential cosinor for graphical analysis as matrix stored in
a list.
}
\description{
Formulas for creating the area of the ellipse to identify
confidence intervals, direction, and graphing purposes.
}
