% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-fit.R
\name{cosinor_features}
\alias{cosinor_features}
\title{Multiple Component Cosinor Features}
\usage{
cosinor_features(object, ...)
}
\arguments{
\item{object}{Model of class \code{cosinor} with multiple periods}

\item{...}{For extensibility}
}
\value{
List of features
}
\description{
Extract the special/global features of a multiple component cosinor. In a multiple component model, there are specific parameters that are not within the model itself, but must be extracted from the model fit. When extracted, can be used to improve the plot of a multiple component cosinor. However, this is only possible if the cosinor is harmonic (see \code{details}). For single-component models, the orthophase is the same as the acrophase and the global amplitude
\itemize{
\item Global Amplitude (Ag) = the overall amplitude is defined as half the difference between the peak and trough values
\item Orthophase (Po) = the lag until the peak time
\item Bathyphase (Pb) =  the lag until the trough time
}
}
\details{
These calculations can only occur if the periods of the cosinor are harmonic - as in, the longest period is a integer multiple of the smallest period (known as the fundamental frequency). Otherwise, these statistics are not accurate or interpretable.
}
