% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-diagnostics.R
\name{cosinor_zero_amplitude}
\alias{cosinor_zero_amplitude}
\title{Zero Amplitude Test}
\usage{
cosinor_zero_amplitude(object, level = 0.95)
}
\arguments{
\item{object}{model of class \code{cosinor}}

\item{level}{confidence level}
}
\value{
Returns a list of test statistics, as well prints out a report of
analysis.
}
\description{
Zero amplitude test assesses how well the circadian pattern fits
the data, essentially detecting the present of a rhythm to the data.
}
