% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrv_modeling.R
\name{geom_residuals}
\alias{geom_residuals}
\title{Plotting Residual of a Model}
\usage{
geom_residuals(model)
}
\arguments{
\item{model}{Model to be analyzed, currently only accepts linear models.}
}
\value{
Returns a ggplot object of geom type, other layers can be added on as seen in example.
}
\description{
`geom_residuals` makes a diagnostic plot of residuals versus fitted data for linear models
}
\details{
Generate residuals versus fitted plot. Functions as an additional geom layer on ggplot
}
\examples{
ggplot() + geom_residuals(model)
}
