% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{get_BFM}
\alias{get_BFM}
\title{baseflow_regime_magnitude}
\usage{
get_BFM(BFA)
}
\arguments{
\item{BF}{Baseflow vector}

\item{days}{Vector containting the days of the (hydrological) year}

\item{n}{Window size for the rolling mean function}
}
\value{
res
}
\description{
Computes the baseflow magnitude: it is the relative
difference between the maxima and minima of the 'n'-days smoothed
inter-annual daily average of baseflow time series. Note in this
second version, the min and the max are also returned and the
default smoothing is 1 (i.e. no smoothing)
}
