% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-constructor.R
\name{ggcosinor}
\alias{ggcosinor}
\title{ggplot of cosinor model}
\usage{
ggcosinor(object, labels = TRUE, ...)
}
\arguments{
\item{object}{Model of class \code{cosinor}.}

\item{labels}{Logical value if annotations should be placed on plot, default
= TRUE. If using a multiple cosinor, the terms are different and will not
be plotted.
\itemize{
\item M = midline estimating statistic of rhythm
\item A = amplitude
\item P = phi or acrophase (shift from 0 to crest)
}}

\item{...}{For extensibility. This function will use different
implementations based on the type of model (single or multiple component).
Attributes of the object will be passed down, or calculated on the fly.}
}
\value{
Object of class \code{ggplot} that can be layered
}
\description{
ggplot of cosinor model that can visualize a variety of cosinor
model subtypes, including single-component, multiple-component, individual,
and population cosinor models, built using \link[card:cosinor]{card::cosinor}.
}
