% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-diagnostics.R
\name{ggcosinor}
\alias{ggcosinor}
\title{ggplot of cosinor model}
\usage{
ggcosinor(object, ...)
}
\arguments{
\item{object}{Model of class \code{cosinor}c("#FFFFFF", "#FFFFFF", "#FFFFFF")}

\item{...}{For extensibility}
}
\description{
ggplot of cosinor model
}
