% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-diagnostics.R
\name{ggcosinor}
\alias{ggcosinor}
\title{ggplot of cosinor model}
\usage{
ggcosinor(object, residuals = TRUE, ...)
}
\arguments{
\item{object}{Model of class \code{cosinor}.}

\item{residuals}{If residuals should be added. Colors can be overwritten with additional arguments (e.g. "+ scale_color_grey()")}

\item{...}{For extensibility}
}
\description{
ggplot of single cosinor model, options include to ability to add diagnostic/fit parameters such as residuals.
}
