% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinor-diagnostics.R
\name{ggcosinorfit}
\alias{ggcosinorfit}
\title{Graphical Assessment of Amplitude and Acrophase}
\usage{
ggcosinorfit(model, alpha = 0.05, ...)
}
\arguments{
\item{model}{Requires a cosinor model to extract the correct statistics to
generate the plot.}

\item{alpha}{Defaults to 0.05.}

\item{...}{Additional parameters may be needed for extensibility}
}
\value{
ggplot object that has has multiple components
}
\description{
This is a ggplot-styled graphical representation of the ellipse
region generated by the cosinor analysis. It requires the same data used by
cosinor model to be fit with the model \link[card:cosinor]{card::cosinor}. This includes
the amplitude, acrophase,
}
\examples{
data("twins")
m <- cosinor(rDYX ~ hour, twins)
ggcosinorfit(m)
}
