% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_dynamic.R
\name{hsaCumsum}
\alias{hsaCumsum}
\title{hsaCumsum}
\usage{
hsaCumsum(x, na.action = "interpolate")
}
\arguments{
\item{x}{vector to use to compute cumulated values}

\item{na.action}{action to undertake when missing values are found
(default: 'interpolate', any other value leads to the original
cumsum() function)}
}
\value{
res
}
\description{
Wrapper around cumsum() to linearly interpolate
missing values
}
