% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matlab_cardiovascular_signal_toolbox.R
\name{read_hrv_matlab}
\alias{read_hrv_matlab}
\title{Read in Toolbox HRV}
\usage{
read_hrv_matlab(loc, name)
}
\arguments{
\item{loc}{Location of the folder that contains all of the patients
that were analyzed by the Main_HRV_Analysis.m function from the
Toolbox}

\item{name}{Name of the patient/ID. There should exist a folder with
the name inside the \code{loc} folder. Inside this folder are all the
Toolbox parameters and HRV results in CSV format.
3600 seconds (which is 1 hour)}
}
\value{
Data frame of HRV summarized by the grouping variable (e.g.
3600 seconds = 1 hour). Also returns an additional column of
percent missing (e.g. 20.0\% missing data) by time group.
}
\description{
\code{read_hrv_matlab} takes the output from HRV Toolbox and
reads it in for an individual patient. Unlike
\code{\link{proc_hrv_matlab}}, this does not process or summarize
the data, it just reads it in raw analysis. Uses the package
\code{\link[data.table:fread]{data.table::fread()}} for reading in data due to size/speed.
}
\details{
The data is taken sequentially (sliding windows), and is not processed. It is reported and is likely a large file.
}
\examples{
# Assuming in "root/code" folder, need to get to data in STUDY
parent <- dirname(getwd())
loc <- file.path(parent, 'data', 'proc_hrv', 'STUDY')
vars <- Hmisc::Cs(SDNN, RMSSD, pnn50, ulf, vlf, lf, hf, lfhf, ttlpwr, ac, dc, SampEn, ApEn)
time <- 3600 # number of seconds in an hour
names <- list.files(path = loc)

# Create for loop to eventually make data frame
d <- list()
for(i in 1:length(names)) {
d[[paste(names[i])]] <- read_hrv_matlab(loc, names[i], vars, time)
}
df <- rbindlist(d) \%>\% as_tibble()
}
