% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetics.R
\name{read_vep_data}
\alias{read_vep_data}
\title{Read VEP Data}
\usage{
read_vep_data(file, columns = NULL)
}
\value{
A \code{tibble} containing the relevant columns from a VCF file that has been annotated by \code{vep} from \href{https://grch37.ensembl.org/info/docs/tools/vep/index.html}{Ensembl-VEP}. The columns can be selected
}
\description{
Read VEP Data
}
