% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent-events.R
\name{recur_summary}
\alias{recur_summary}
\title{Recurrent Event Summary Table by Group}
\usage{
recur_summary(data, covar)
}
\arguments{
\item{data}{Recurrent event data in marginal format. There must be
an ID column. Must merge in the covariate of interest into this data set.}

\item{covar}{Name of covariate of interest to serve as grouping variable.}
}
\value{
Summary table by grouping variable, can be placed into a latex
environment with kableExtra. Assumes that death events may be
present when most recent non-EVENT has status 1.
}
\description{
\code{recur_summary} Creates a table with summary of
recurrent events
}
\details{
This function allows for taking the output of
\link[card:recur_survival_table]{card::recur_survival_table} marginal format repeat event data,
and creates a summary table that describes the number of events by
strata/event.
}
\examples{
# setup
library(magrittr)
library(dplyr)
library(kableExtra)

data("mims")

# Marginal data (as above)
marg <-
  recur_survival_table(
    mims, "patid", "first_visit_date_bl", "ldka",
    c("mi_date_1", "mi_date_2", "mi_date_3"), "marginal", "DEATH_CV_YN"
  )

# Grouping variable
covar <- "RDR_ISCHEMIA_M_YN_bl"
df <- left_join(marg, mims[c("patid", covar)], by = c("ID" = "patid"))

# Make summary table
tbl <- recur_summary(df, covar)
tbl \%>\%
  kable("latex", caption = "Title", booktabs = TRUE) \%>\%
  kable_styling(font_size = 8)
}
