% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent_survival_data.R
\name{recurrent_propensity}
\alias{recurrent_propensity}
\title{Propensity Score Weighting}
\usage{
recurrent_propensity(data, vars)
}
\arguments{
\item{data}{Data frame that contains all covariates and outcomes. First column should be ID}

\item{vars}{Variables used for regression. Outcome variable must be first.}
}
\value{
Returns a modified table from what was originally given with the new columns propensity scores. Essentially original df + 2 columns.
}
\description{
\code{recurrent_propensity} Adds propensity score to any data set that is being regressed upon.
}
\details{
Using a logistic regression, will take covariates and create propensity scores, and adds the weights. Uses the standard logistic regression to evalute the propensity score.
}
\examples{
ps <- recurrent_propensity(df, c("outcome", "exp1", "exp2"))
ps[c("patid", "PROP_SCORE", "PROP_WEIGHT")]
}
