% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent_survival_data.R
\name{recurrent_survival_table}
\alias{recurrent_survival_table}
\title{Recurrent Survival Data Format}
\usage{
recurrent_survival_table(
  data,
  id,
  first,
  last,
  event.dates,
  model.type,
  death = NULL
)
}
\arguments{
\item{data}{A dataframe containing the subsequent parameters}

\item{id}{Column in dataframe that contains unique IDs for each row}

\item{first}{Column with left/enrollment dates}

\item{last}{Column with right/censoring time point, or last contact}

\item{event.dates}{Vector of columns that contain event dates}

\item{model.type}{Character/string = c("marginal", "pwptt", "pwpgt")}

\item{death}{Column created if death is known (0 or 1), Must include in original dataframe (e.g. can add column of zeroes prn)}
}
\value{
A data frame organized into a survival table format
}
\description{
\code{recurrent_survival_table} Reformat recurrent event data (wide) into different models for survival analysis
}
\details{
This function takes every data event date, and creates several types of recurrent event tables. It orders the data chronologically for repeat events. Currently does marginal and conditional A and B models.
}
\examples{
recurrent_survival_table(tibble, id, first.date, last.known.date, c(event.dates), "marginal", death.status)
}
