% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recurrent_create_table.R
\name{recurrent_survival_table}
\alias{recurrent_survival_table}
\title{Create data table of recurrent events for survival analysis}
\usage{
recurrent_survival_table(data, id, first, last, event.dates, model.type, death)
}
\arguments{
\item{data}{A dataframe containing the subsequent parameters}

\item{id}{Column in dataframe that contains unique IDs for each row}

\item{first}{Column with left/enrollment dates}

\item{last}{Column with right/censoring time point, or last contact}

\item{event.dates}{Vector of columns that contain event dates}

\item{model.type}{Character/string = c("marginal", "pwptt", "pwpgt")}

\item{death}{Column representing if patient died at right censoring (optional), default = 0 (no death event)}
}
\value{
A data frame organized into a survival table format
}
\description{
Create data table of recurrent events for survival analysis
}
\examples{
recurrent_survival_table(tibble, id, first.date, last.known.date, c(event.dates), "marginal", death.status)
}
