% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_dynamic.R
\name{rq_slopes}
\alias{rq_slopes}
\title{pq_slopes}
\usage{
rq_slopes(
  Q,
  R,
  hdays,
  start = 15,
  end = 183,
  bp = mean(c(start, end)),
  intercept = TRUE
)
}
\arguments{
\item{Q}{Streamflow vector}

\item{R}{precipitation vector}

\item{hdays}{Days of the (Hydrological) years vector}

\item{start}{start of period to search for change of trend in R-Q}

\item{end}{end of period to search for change of trend in R-Q}

\item{bp}{initial guess of threshold date}

\item{intercept}{should the intercept be estimated (default: TRUE)
or fixed to c(0, 0) (FALSE) ?}
}
\value{
res
}
\description{
Computes the cumulative inter-annual daily average of
R and Q as well as the difference R-Q. Then, compute the seasonal
response change signatures following the so-called R-Q approach :
breackpoint date, first period slope (dry), second period slope
(wet) and the intercepts associated with these two slopes (only
there for plotting purposes).
}
