% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maude-adjudicate.R
\name{validate_complication_index}
\alias{validate_complication_index}
\title{Internal helper to validate that a complication index is a
named list of character vectors whose names are present in the supplied
complication definitions, with optional \code{"not_indexed"} allowed.}
\usage{
validate_complication_index(index, index_names)
}
\description{
Internal helper to validate that a complication index is a
named list of character vectors whose names are present in the supplied
complication definitions, with optional \code{"not_indexed"} allowed.
}
\keyword{internal}
